+++ application/config.py
@@ -1,0 +1,4 @@
+class Config:
+    DEBUG = True
+    DB_URL = "mongodb://localhost:27017"
+    DB_NAME = "web-forum"
